Refactor DTO Python renderer
